requirements.txt
gradio
watchdog
psutil
Pillow
numpy
ipfshttpclient
shutil
